requirements.txt
autopep8==2.3.1
certifi==2024.8.30
chardet==5.2.0
defusedxml==0.7.1
Django==5.1.1
django-allauth==64.2.1
django-countries==7.6.1
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
idna==3.10
oauthlib==3.2.2
pep8==1.7.1
Pillow==10.4.0
pycodestyle==2.12.1
python-decouple==3.8
stripe==10.11.0
urllib3==2.2.3
django-multiselectfield==0.1.13
django-breadcrumbs==1.1.3
scipy==1.14.1
svg.path
potracer==0.0.4
tqdm
svgwrite==1.4.3
numpy==1.26.1
potrace-cli==0.0.4
opencv-python==4.10.0.84
scikit-learn==1.3.0
django-redis
waifu2x_ncnn_py
waifu2x-ncnn-vulkan-python
svgpathtools
shapely
ultralytics==8.0.183
django-subscriptions
fastapi
django-rest-auth 
twilio
django-paypal
paypalrestsdk
python-social-auth==0.2.12
alibabacloud_dysmsapi20170525==3.1.0